move llm module related things in llm_generate that deals with generations
