+++ llm/models/__init__.py
@@ -1,4 +1,3 @@
 from . import llm_model
 from . import llm_provider
 from . import llm_publisher
-from . import llm_generation_config
