+++ llm/models/llm_model.py
@@ -44,9 +44,6 @@ class LLMModel(models.Model):
             ("completion", "Completion"),
             ("chat", "Chat"),
             ("multimodal", "Multimodal"),
-            ("image_generation", "Image Generation"),
-            ("audio_generation", "Audio Generation"),
-            ("video_generation", "Video Generation"),
         ]
 
     @api.model_create_multi
@@ -87,95 +84,3 @@ class LLMModel(models.Model):
             },
         }
     
-    def _is_generative_task_model(self):
-        """ Helper to check if model_use indicates a non-chat/embedding generative task. """
-        self.ensure_one()
-        return self.model_use in ['image_generation', 'audio_generation', 'video_generation']
-        
-    def action_generate_llm_generation_config(self):
-        """Generate a generation configuration from the model's details.
-        This reads from self.details and dispatches to the provider.
-        """
-        self.ensure_one()
-            
-        # Dispatch to provider-specific implementation
-        self.provider_id._dispatch(
-            "get_config_from_raw_schema",
-            model_record=self
-        )
-
-        return True
-    
-    def action_generate_media(self):
-        self.ensure_one()
-        
-        # Check if we have a generation config
-        if not self.generation_config_id and self._is_generative_task_model():
-            return self.action_generate_llm_generation_config()
-        
-        # Prepare inputs based on the model type
-        if self._is_generative_task_model():
-            # For media generation models, we need to structure the input according to Replicate's API
-            inputs = {
-                "prompt": "A beautiful sunset over mountains"
-            }
-        else:
-            raise ValueError(f"Model {self.name} is not configured for media generation")
-        
-        # Dispatch to provider-specific implementation
-        result = self.provider_id._dispatch(
-            "generate_media",
-            inputs=inputs,
-            model_record=self,
-            stream=False
-        )
-        
-        # Log the result
-        return result
-    
-    def generate_media(self, inputs, stream=False):
-        """Generate content using this model with the specified inputs.
-        
-        Args:
-            inputs (dict): The input parameters for generation according to the schema
-            
-        Returns:
-            The generated content (format depends on the model type and configuration)
-        """
-        self.ensure_one()
-        
-        # Validate model is configured for generation
-        if not self._is_generative_task_model():
-            raise ValueError(f"Model {self.name} is not configured for generation tasks")
-            
-        if not self.generation_config_id:
-            raise ValueError(f"Model {self.name} requires a generation configuration")
-        
-        # Dispatch to provider-specific implementation
-        return self.provider_id._dispatch(
-            "generate_media",
-            inputs=inputs,
-            model_record=self,
-            stream=stream
-        )
-
-    def format_generation_response(self, raw_response):
-        """Format the raw generation response according to the output processing config
-        
-        Args:
-            raw_response: The raw response from the provider
-            
-        Returns:
-            Processed response in the format specified by the config
-        """
-        self.ensure_one()
-        
-        if not self.generation_config_id:
-            raise ValueError(f"Model {self.name} requires a generation configuration")
-        
-        # Dispatch to provider-specific implementation
-        return self.provider_id._dispatch(
-            "format_generation_response",
-            raw_response=raw_response,
-            output_schema=self.generation_config_id.output_schema_raw
-        )
