+++ llm_generate/security/ir.mode.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_llm_generation_config_user,llm.generation.config.user,model_llm_generation_config,base.group_user,1,0,0,0
+access_llm_generation_config_manager,llm.generation.config.manager,model_llm_generation_config,group_llm_manager,1,1,1,1
