+++ llm_generate/views/llm_menu_views.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <!-- Menu Item -->
+    <menuitem
+    id="menu_llm_generation_config"
+    name="Generation Configs"
+    parent="menu_llm_config"
+    action="llm_generation_config_action"
+    sequence="30"
+    />
+</odoo>
